+++ client/src/Extension.ts
@@ -1,19 +1,26 @@
 import * as path from 'path'
 import {
-	workspace as Workspace, window as Window, ExtensionContext, TextDocument, OutputChannel, WorkspaceFolder, Uri, workspace, commands
+	ExtensionContext,
+	LanguageStatusItem,
+	OutputChannel,
+	TextDocument,
+	Uri,
+	window as Window,
+	workspace as Workspace,
+	WorkspaceFolder,
+	commands, languages,
+	workspace
 } from 'vscode'
 
 import {
 	LanguageClient, LanguageClientOptions, ServerOptions, TransportKind
 } from 'vscode-languageclient/node'
 import { NeosContextStatusBarItem, NeosContextStatusBarItemClass } from './neosContextStatusBarItem'
+import { PreferenceService } from './PreferenceService'
+import { ProgressNotificationService } from './ProgressNotificationService'
 import { AbstractCommandConstructor } from './commands/AbstractCommand'
 import { InspectCommand } from './commands/InspectCommand'
 import { ReloadCommand } from './commands/ReloadCommand'
-import { PreferenceService } from './PreferenceService'
-import { ProgressNotificationService } from './ProgressNotificationService'
-import { StatusItemService } from './StatusItemService'
-
 
 
 export class Extension {
@@ -22,10 +29,28 @@ export class Extension {
 	protected sortedWorkspaceFolders: string[] | undefined = undefined
 	protected context: ExtensionContext | undefined = undefined
 
+	protected languageStatusBarItems: {
+		reload: LanguageStatusItem
+	} = { reload: undefined }
+
 	constructor() {
 		this.outputChannel = Window.createOutputChannel('Neos Fusion LSP')
 
 		Workspace.onDidChangeWorkspaceFolders(() => this.sortedWorkspaceFolders = undefined)
+
+		this.createLanguageStatusItems()
+	}
+
+	protected createLanguageStatusItems() {
+		const documentSelector = { scheme: 'file', language: 'fusion' }
+		this.languageStatusBarItems.reload = languages.createLanguageStatusItem("fusion.reload", documentSelector)
+		this.languageStatusBarItems.reload.name = "reload"
+		this.languageStatusBarItems.reload.text = "Reload Fusion language server"
+		this.languageStatusBarItems.reload.command = {
+			title: "reload",
+			command: "neos-fusion-lsp.reload",
+			tooltip: "Reload the Fusion Language Server"
+		}
 	}
 
 	public getClients() {
@@ -61,7 +86,7 @@ export class Extension {
 	}
 
 	protected onDidOpenTextDocument(document: TextDocument) {
-		if (document.languageId !== 'fusion' || (document.uri.scheme !== 'file' && document.uri.scheme !== 'untitled')) return
+		if (document.uri.scheme !== 'file' && document.uri.scheme !== 'untitled') return
 
 		const uri = document.uri
 		const folder = Workspace.getWorkspaceFolder(uri)
@@ -70,9 +95,8 @@ export class Extension {
 		const outerMostWorkspaceFolder = this.getOuterMostWorkspaceFolder(folder)
 		if (this.clients.has(outerMostWorkspaceFolder.uri.toString())) return
 
-		const inspect = workspace.getConfiguration().get("neosFusionLsp.logging.inspect", false)
-
-		const startedClient = this.startClient(outerMostWorkspaceFolder, inspect)
+		const startClientInInspectMode = workspace.getConfiguration().get("neosFusionLsp.logging.inspect", false)
+		const startedClient = this.startClient(outerMostWorkspaceFolder, startClientInInspectMode)
 
 		NeosContextStatusBarItem.init(this.context, startedClient, this.outputChannel)
 
@@ -141,7 +165,8 @@ export class Extension {
 				fileEvents: [
 					workspace.createFileSystemWatcher('**/*.php'),
 					workspace.createFileSystemWatcher('**/*.yaml'),
-					workspace.createFileSystemWatcher('**/*.fusion')
+					workspace.createFileSystemWatcher('**/*.fusion'),
+					workspace.createFileSystemWatcher('**/*.yaml')
 				]
 			},
 			initializationOptions: {
@@ -151,16 +176,15 @@ export class Extension {
 			}
 		}
 
-		const statusItemService = new StatusItemService(documentSelector)
 		const progressNotificationService = new ProgressNotificationService()
 		const client = new LanguageClient('vscode-neos-fusion-lsp', 'LSP For Neos Fusion (and AFX)', serverOptions, clientOptions)
 
-		client.onNotification('custom/busy/create', ({ id, configuration }) => statusItemService.createStatusItem(id, configuration))
+		client.onNotification('custom/busy/create', () => this.languageStatusBarItems.reload.busy = true)
 		client.onNotification('custom/progressNotification/create', ({ id, title }) => progressNotificationService.create(id, title))
 
 		client.onNotification('custom/progressNotification/update', ({ id, payload }) => progressNotificationService.update(id, payload))
 
-		client.onNotification('custom/busy/dispose', ({ id }) => statusItemService.disposeStatusItem(id))
+		client.onNotification('custom/busy/dispose', () => this.languageStatusBarItems.reload.busy = false)
 		client.onNotification('custom/progressNotification/finish', ({ id }) => progressNotificationService.finish(id))
 
 		client.start()
@@ -169,7 +193,7 @@ export class Extension {
 		return client
 	}
 
-	public stopClients() {
+	public async stopClients() {
 		const promises: Thenable<void>[] = []
 
 		for (const client of this.clients.values()) {