+++ server/src/neos/NeosPackage.ts
@@ -46,6 +46,10 @@ export class NeosPackage extends Logger {
 		this.debug = this.getName() === "neos/fusion"
 
 		this.initNamespaceLoading()
+		this.readConfiguration()
+	}
+
+	public readConfiguration() {
 		this.configuration = FlowConfiguration.ForPackage(this)
 	}